requirements.txt
flask
flask-cors
flask-socketio
flask-session
psycopg2-binary
gunicorn
python-json-logger
python-logging-loki
alembic
redis
eventlet
